+++ limitless_sdk/orders/__init__.py
@@ -3,6 +3,7 @@
 from .builder import OrderBuilder
 from .signer import OrderSigner
 from .client import OrderClient
+from .receive_window import normalize_receive_window_options
 from .validator import (
     ValidationError,
     validate_gtc_order_args,
@@ -15,6 +16,7 @@ __all__ = [
     "OrderBuilder",
     "OrderSigner",
     "OrderClient",
+    "normalize_receive_window_options",
     "ValidationError",
     "validate_gtc_order_args",
     "validate_fok_order_args",